Script

+++ uninstall.sh
@@ -7,6 +7,7 @@ cd "$here"
 for f in .??*
 do
   test "$f" = .git && continue
+  test "$f" = .github && continue
   test "$f" = .gitignore && continue
   unlink "$HOME/$f"
 done
